feature/GH-1178: create guest register form component
