+++ projects/storefrontlib/src/cms-components/order-confirmation/components/order-confirmation-thank-you-message/order-confirmation-thank-you-message.component.spec.ts
@@ -1,15 +1,6 @@
-import { Component, Type, DebugElement } from '@angular/core';
+import { Component, Input, Type } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { ReactiveFormsModule } from '@angular/forms';
-import {
-  CheckoutService,
-  I18nTestingModule,
-  Order,
-  UserToken,
-  AuthService,
-  UserService,
-  RoutingService,
-} from '@spartacus/core';
+import { CheckoutService, I18nTestingModule, Order } from '@spartacus/core';
 import { Observable, of } from 'rxjs';
 import { OrderConfirmationThankYouMessageComponent } from './order-confirmation-thank-you-message.component';
 
@@ -18,6 +9,11 @@ import { By } from '@angular/platform-browser';
 
 @Component({ selector: 'cx-add-to-home-screen-banner', template: '' })
 class MockAddtoHomeScreenBannerComponent {}
+
+@Component({ selector: 'cx-guest-register-form', template: '' })
+class MockGuestRegisterFormComponent {
+  @Input() guid;
+}
 
 class MockCheckoutService {
   clearCheckoutData = createSpy();
@@ -31,42 +27,21 @@ class MockCheckoutService {
   }
 }
 
-class MockAuthService {
-  getUserToken(): Observable<UserToken> {
-    return of({ access_token: 'test' } as UserToken);
-  }
-}
-
-class MockUserService {
-  registerGuest = createSpy();
-}
-
-class MockRoutingService {
-  go = jasmine.createSpy('go');
-}
-
 describe('OrderConfirmationComponent', () => {
   let component: OrderConfirmationThankYouMessageComponent;
   let fixture: ComponentFixture<OrderConfirmationThankYouMessageComponent>;
-  let form: DebugElement;
 
-  let userService: UserService;
   let checkoutService: CheckoutService;
-  let routingService: RoutingService;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [I18nTestingModule, ReactiveFormsModule],
+      imports: [I18nTestingModule],
       declarations: [
         OrderConfirmationThankYouMessageComponent,
         MockAddtoHomeScreenBannerComponent,
+        MockGuestRegisterFormComponent,
       ],
-      providers: [
-        { provide: CheckoutService, useClass: MockCheckoutService },
-        { provide: AuthService, useClass: MockAuthService },
-        { provide: UserService, useClass: MockUserService },
-        { provide: RoutingService, useClass: MockRoutingService },
-      ],
+      providers: [{ provide: CheckoutService, useClass: MockCheckoutService }],
     }).compileComponents();
   }));
 
@@ -74,13 +49,8 @@ describe('OrderConfirmationComponent', () => {
     fixture = TestBed.createComponent(
       OrderConfirmationThankYouMessageComponent
     );
-
-    userService = TestBed.get(UserService as Type<UserService>);
-    checkoutService = TestBed.get(CheckoutService as Type<CheckoutService>);
-    routingService = TestBed.get(RoutingService as Type<RoutingService>);
-
     component = fixture.componentInstance;
-    form = fixture.debugElement.query(By.css('form'));
+    checkoutService = TestBed.get(CheckoutService as Type<CheckoutService>);
   });
 
   it('should create', () => {
@@ -88,50 +58,33 @@ describe('OrderConfirmationComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should do nothing when submit from non-guest customer', () => {
+  it('should display order code', () => {
+    component.ngOnInit();
+    fixture.detectChanges();
+    expect(
+      fixture.debugElement.query(By.css('.cx-page-title')).nativeElement
+        .innerHTML
+    ).toContain('test-code-412');
+  });
+
+  it('should display guest register form for guest user', () => {
+    component.ngOnInit();
+    fixture.detectChanges();
+
+    expect(
+      fixture.debugElement.query(By.css('cx-guest-register-form'))
+    ).not.toBeNull();
+  });
+
+  it('should not display guest register form for login user', () => {
     spyOn(checkoutService, 'getOrderDetails').and.returnValue(
       of({ guid: 'guid', guestCustomer: false })
     );
     component.ngOnInit();
     fixture.detectChanges();
-    component.submit();
 
-    expect(userService.registerGuest).not.toHaveBeenCalledWith();
-  });
-
-  it('should register customer and redirect to homepage when submit from guest customer', () => {
-    const password = 'test password';
-    component.guestRegisterForm.controls['password'].setValue(password);
-    component.ngOnInit();
-    fixture.detectChanges();
-    component.submit();
-
-    expect(userService.registerGuest).toHaveBeenCalledWith('guid', password);
-    expect(routingService.go).toHaveBeenCalledWith({ cxRoute: 'home' });
-  });
-
-  describe('UI test', () => {
-    it('should display order code', () => {
-      component.ngOnInit();
-      fixture.detectChanges();
-      expect(
-        fixture.debugElement.query(By.css('.cx-page-title')).nativeElement
-          .innerHTML
-      ).toContain('test-code-412');
-    });
-
-    it('shoud not have form for login user', () => {
-      spyOn(checkoutService, 'getOrderDetails').and.returnValue(
-        of({ guid: 'guid', guestCustomer: false })
-      );
-      component.ngOnInit();
-      fixture.detectChanges();
-
-      expect(form).toEqual(null);
-    });
-
-    it('shoud thankYou message contain email', () => {
-      // need UI work done first
-    });
+    expect(
+      fixture.debugElement.query(By.css('cx-guest-register-form'))
+    ).toBeNull();
   });
 });
