+++ projects/storefrontlib/src/cms-components/order-confirmation/components/order-confirmation-thank-you-message/order-confirmation-thank-you-message.component.ts
@@ -4,17 +4,9 @@ import {
   OnDestroy,
   OnInit,
 } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import {
-  CheckoutService,
-  Order,
-  UserService,
-  RoutingService,
-  AuthService,
-} from '@spartacus/core';
-import { Observable, Subscription } from 'rxjs';
+import { CheckoutService, Order } from '@spartacus/core';
+import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
-import { CustomFormValidators } from '../../../../shared/utils/validators/custom-form-validators';
 
 @Component({
   selector: 'cx-order-confirmation-thank-you-message',
@@ -27,39 +19,7 @@ export class OrderConfirmationThankYouMessageComponent
   isGuestCustomer = false;
   orderGuid: string;
 
-  subscription: Subscription;
-
-  guestRegisterForm: FormGroup = this.fb.group(
-    {
-      password: [
-        '',
-        [Validators.required, CustomFormValidators.passwordValidator],
-      ],
-      passwordconf: ['', Validators.required],
-    },
-    { validator: CustomFormValidators.matchPassword }
-  );
-
-  constructor(
-    checkoutService: CheckoutService,
-    userService: UserService, // tslint:disable-line
-    routingService: RoutingService, // tslint:disable-line
-    authService: AuthService, // tslint:disable-line
-    fb: FormBuilder // tslint:disable-line
-  );
-  /**
-   * @deprecated since version 1.x
-   *
-   *  TODO(issue:#1178) deprecated since 1.x
-   */
-  constructor(checkoutService: CheckoutService);
-  constructor(
-    protected checkoutService: CheckoutService,
-    protected userService?: UserService,
-    protected routingService?: RoutingService,
-    protected authService?: AuthService,
-    protected fb?: FormBuilder
-  ) {}
+  constructor(protected checkoutService: CheckoutService) {}
 
   ngOnInit() {
     this.order$ = this.checkoutService.getOrderDetails().pipe(
@@ -72,24 +32,5 @@ export class OrderConfirmationThankYouMessageComponent
 
   ngOnDestroy() {
     this.checkoutService.clearCheckoutData();
-    if (this.subscription) {
-      this.subscription.unsubscribe();
-    }
-  }
-
-  submit() {
-    if (this.isGuestCustomer) {
-      this.userService.registerGuest(
-        this.orderGuid,
-        this.guestRegisterForm.value.password
-      );
-      if (!this.subscription) {
-        this.subscription = this.authService.getUserToken().subscribe(token => {
-          if (token.access_token) {
-            this.routingService.go({ cxRoute: 'home' });
-          }
-        });
-      }
-    }
   }
 }
